Move type-dispatcher calls from traits.hpp to traits.cpp
