+++ cpp/src/hash/md5_hash.cu
@@ -208,7 +208,7 @@ struct ListHasherDispatcher {
 };
 
 // MD5 supported leaf data type check
-constexpr inline bool md5_leaf_type_check(data_type dt)
+inline bool md5_leaf_type_check(data_type dt)
 {
   return (is_fixed_width(dt) && !is_chrono(dt)) || (dt.id() == type_id::STRING);
 }
