+++ cpp/src/io/statistics/column_statistics.cuh
@@ -101,28 +101,33 @@ struct calculate_group_statistics_functor {
                               !std::is_same_v<T, list_view>)>* = nullptr>
   __device__ void operator()(stats_state_s& s, uint32_t t)
   {
-    detail::storage_wrapper<block_size> storage(temp_storage);
+    // Temporarily disable stats writing for int96 timestamps
+    // TODO: https://github.com/rapidsai/cudf/issues/10438
+    if constexpr (!cudf::is_timestamp<T>() or IO != detail::io_file_format::PARQUET or
+                  INT96 != detail::is_int96_timestamp::YES) {
+      detail::storage_wrapper<block_size> storage(temp_storage);
 
-    using type_convert = detail::type_conversion<detail::conversion_map<IO, INT96>>;
-    using CT           = typename type_convert::template type<T>;
-    typed_statistics_chunk<CT, detail::statistics_type_category<T, IO>::include_aggregate> chunk;
+      using type_convert = detail::type_conversion<detail::conversion_map<IO, INT96>>;
+      using CT           = typename type_convert::template type<T>;
+      typed_statistics_chunk<CT, detail::statistics_type_category<T, IO>::include_aggregate> chunk;
 
-    for (uint32_t i = 0; i < s.group.num_rows; i += block_size) {
-      uint32_t r   = i + t;
-      uint32_t row = r + s.group.start_row;
-      if (r < s.group.num_rows) {
-        if (s.col.leaf_column->is_valid(row)) {
-          auto converted_value = type_convert::convert(get_element<T>(s, row));
-          chunk.reduce(converted_value);
-        } else {
-          chunk.null_count++;
+      for (uint32_t i = 0; i < s.group.num_rows; i += block_size) {
+        uint32_t r   = i + t;
+        uint32_t row = r + s.group.start_row;
+        if (r < s.group.num_rows) {
+          if (s.col.leaf_column->is_valid(row)) {
+            auto converted_value = type_convert::convert(get_element<T>(s, row));
+            chunk.reduce(converted_value);
+          } else {
+            chunk.null_count++;
+          }
         }
       }
+
+      chunk = block_reduce(chunk, storage);
+
+      if (t == 0) { s.ck = get_untyped_chunk(chunk); }
     }
-
-    chunk = block_reduce(chunk, storage);
-
-    if (t == 0) { s.ck = get_untyped_chunk(chunk); }
   }
 
   template <typename T,
@@ -303,11 +308,13 @@ __global__ void __launch_bounds__(block_size, 1)
     }
     // Temporarily disable stats writing for int96 timestamps
     // TODO: https://github.com/rapidsai/cudf/issues/10438
-    else if (not cudf::is_timestamp(state.col.leaf_column->type())) {
-      type_dispatcher(state.col.leaf_column->type(),
-                      calculate_group_statistics_functor<block_size, IO>(storage),
-                      state,
-                      threadIdx.x);
+    else {
+      type_dispatcher(
+        state.col.leaf_column->type(),
+        calculate_group_statistics_functor<block_size, IO, detail::is_int96_timestamp::YES>(
+          storage),
+        state,
+        threadIdx.x);
     }
   } else {
     type_dispatcher(state.col.leaf_column->type(),
